refactor(Chat): move auto-scroll from MessageList to Layout
XDSChatMessageList is now purely presentational — messages, density
context, empty state, spacer, infinite scroll sentinel. No scroll
behavior, no scroll button.

XDSChatLayout owns all scroll concerns:
- useAutoScroll for scroll-to-bottom and new message detection
- ResizeObserver on the message content area (via contentRef in
  context) to detect height changes from new/streaming content
- ScrollToBottomButton rendered above the composer dock
- Scroll listener attached to the scroll container

The layout context provides contentRef — the message list attaches
it to its inner element so the layout can observe height changes.
No effects on children, no callbacks — just measurement.

useAutoScroll and ScrollToBottomButton remain exported for
standalone composition without XDSChatLayout.
