+++ packages/core/src/Chat/XDSChatContext.tsx
@@ -64,6 +64,8 @@ export interface XDSChatLayoutContextValue {
   scrollContainerRef: React.RefObject<HTMLElement | null>;
   /** Height of the composer dock in px. Used to offset the scroll-to-bottom button. */
   dockHeight: number;
+  /** Ref callback for the message list content area — layout observes it for height changes. */
+  contentRef: (el: HTMLElement | null) => void;
 }
 
 export const XDSChatLayoutContext =
